move tuple to pkg folder to conform to file location standards
