+++ pkg/tuple/pick.go
